Moved toolbox section
I think this repo is an excellent idea! I would only suggest one change to the structure. Since the repo is aimed at developers and/or students who are interesting in getting started with FDNs, I think the "Toolboxes and libraries" section should be the first thing they see: they're probably eager to get their hands dirty, more than reading up on the state of the art.

Similarly, do you think there could be a section for "Open-source implementations"? All of the toolboxes and libraries so far are in MATLAB or Python; audio developers will probably be looking for runtime code. I am less knowledgeable in this regard, but for starters, I can say that Josh's plugin (github.com/jmannall/RoomAcoustiCpp) includes a C++ FDN.
I wouldn't include it among "Toolboxes and libraries" though, because the FDN is just a tiny part of the plugin, which is definitely not intended to be a "tool for FDN development".